Create README.md
During my internship at Cognifyz, I worked on a comprehensive data analysis project focusing on a dataset related to restaurants. The objective was to extract valuable insights from the data, perform preprocessing, conduct descriptive and geospatial analysis, and engage in feature engineering to enrich the dataset for further analysis and modeling.